requirement.txt
absl-py==0.7.1
astor==0.8.0
backports.weakref==1.0.post1
decorator==4.4.1
enum34==1.1.6
funcsigs==1.0.2
futures
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-squeezenet==0.4
Markdown==3.1.1
mock==3.0.5
networkx==2.2
nose==1.3.7
numpy==1.16.3
olefile==0.44
opencv-python==4.1.0.25
Pillow==6.2.2
protobuf==3.7.1
pygame==1.9.6
PyMySQL==0.9.3
PyYAML==5.1
qrcode==5.3
reportlab==3.5.42
scikit-learn==0.20.4
scipy==1.2.2
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.4
